Switch to Google Truth library for assertions
The Google Truth library is popular and easy to read. I switched all the
assertions to use it, but I used the "withMessage" style call for some,
and left some with assertThat with no message. It's easy to just say
"withMessage is only used when it's not obvious from context," but it's
hard to say what will be obvious from context.