Add a Login Activity with a ViewModel and tests
ViewModel can be created for tests by mocking the constructor
parameters, or can be instantiated with the custom ViewModel factory.

LoginViewModelTest covers everything except onCleared, which is package
protected.

Login screen allows user to enter an email address and password for
their bitbucket account, and then attempts to log in to that account and
retrieve their user information.