Replace Vault with androidx.security crypto lib that provides encrypted shared prefs.
NOTE: Due to the nature of this reference application (not published in prod), no migration path was added and no cleanup of existing (now obsolete) prefs files was performed. If you had a prior version installed, you'll need to sign in again. Clear data will remove the obsolete prefs files.
**If you have an app in production with Vault, you WILL need to add a migration path and cleanup obsolete files.**