Correct invalid data sent to the bitbucket repository apis
### Observations
After creating a new OAuth consumer key/secret and test bitbucket account to use for testing the app flow and UI, I observed that repositories were not loading properly. Looking further into the api docs, it was mentioned that the workspace slug (id) or uuid should be used instead of passing in the repository owner nickname (which might have worked earlier due to the values being the same for the previous test Bitbucket account used).

### Changes
* Refactor owner to workspace/workspaceSlug starting with the BitbucketService definitions and bubbling up through the call hierarchy
* Send repository.workspace.slug in place of repository.owner.nickname from all appropriate call sites
* Update README documentation mentioning additional detail around Bitbucket OAuth consumer key creation and setup