Add `StringIdHelper` and inject `Clock` to better support testability
### Use `StringIdHelper` to avoid depending on `Context`
Use `StringIdHelper.*` types (representing string resources, format strings, and plurals as well as raw strings) in `ViewModel`s (or other classes) to avoid needing a `Context`. This allows for **simpler unit testing** and **prevents the need for `Context` in places it shouldn't be used** by pushing the resolution of the final String to the view layer (using the `textByStringIdHelper` DataBinding custom BindingAdapter function that takes in a `Context`).

### Inject `Clock`
Pass the `java.time.Clock` instance from Koin to any java 8 date/time apis that can use it in order **to allow control of the clock in unit tests**. This will typically be the `.now(clock: Clock)` api on `Instant`, `ZonedDateTime`, `LocalDateTime`, `LocalDate`, `LocalTime`, and so on.

### DateTimeFormatter change
Using the `DateTimeFormatter.ISO_LOCAL_DATE` formatter instead of `DateTimeFormatter.ISO_DATE` to prevent showing a date with `Z` or `-06:00` in the last updated date UI.

### Additional Unit Tests
Add unit tests for `DateTimeUtils` (controlling the clock for easy testing).

### Documentation
Update `BEST_PRACTICES.md` with entries for `StringIdHelper` and injecting the Clock.